Change max daily volume into a int64 type field

+++ db/trade_summary.go
@@ -3,5 +3,5 @@ package db
 type TradeSummary struct {
 	Ticker         string  `json:"ticker"`
 	MaxRangeValue  float64 `json:"max_range_value"`
-	MaxDailyVolume int     `json:"max_daily_volume"`
+	MaxDailyVolume int64   `json:"max_daily_volume"`
 }
